various minor fixes after testing
