+++ addons/sys_core/initSettings.sqf
@@ -164,5 +164,5 @@
     "LIST",
     localize LSTRING(DefaultZeusVoiceLocation_displayName),
     "ACRE2",
-    [[false, true], ["STR_A3_Leaderboards_Header_Player", "str_a3_cfgvehicles_moduleremotecontrol_f"], 1]
+    [[false, true], ["str_a3_cfgvehicles_moduleremotecontrol_f", "STR_A3_Leaderboards_Header_Player"], 0]
 ] call CBA_Settings_fnc_init;
