+++ addons/sys_zeus/ACE_Actions.hpp
@@ -1,0 +1,17 @@
+class ACE_ZeusActions {
+    class ACRE_ZeusEars {
+        displayName = "ACRE";
+        condition = QUOTE(true);
+        statment = "";
+        class ACRE_SpectatorEars {
+            displayName = CSTRING(spectatorEars);
+            condition = QUOTE(!ACRE_IS_SPECTATOR);
+            statement = QUOTE([true] call EFUNC(api,setSpectator));
+        };
+        class ACRE_ZeusEars {
+            displayName = CSTRING(zeusEars);
+            condition = QUOTE(ACRE_IS_SPECTATOR);
+            statement = QUOTE([false] call EFUNC(api,setSpectator));
+        };
+    };
+};
