+++ addons/sys_zeus/stringtable.xml
@@ -10,7 +10,7 @@
         <Key ID="STR_ACRE_sys_zeus_remote">
             <English>Use Remote Controlled Voice</English>
         </Key>
-        <Key ID="STR_ACRE_sys_zeus_origin">
+        <Key ID="STR_ACRE_sys_zeus_player">
             <English>Use Player Voice</English>
         </Key>
